backend/requirements.txt
Django==2.0.2
djangorestframework==3.7.7
markdown==2.6.11
django-filter==1.1.0